+++ lib/src/io.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'dart:io';
+import 'dart:typed_data';
 
 /// Loads a file as String using [encoding].
 ///
@@ -9,10 +10,20 @@ Future<String> catFile(File file, [Encoding encoding]) async {
   return file.readAsString(encoding: encoding);
 }
 
-/// Saves a file from [data] string using [encoding].
+/// Loads a [file] bytes.
+Future<Uint8List> catFileBytes(File file) async {
+  return file.readAsBytes();
+}
+
+/// Saves a [file] from [data] string using [encoding].
 ///
 /// [encoding] If null uses UTF-8.
 Future<File> saveFile(File file, String data, [Encoding encoding]) async {
   encoding ??= utf8;
   return file.writeAsString(data, encoding: encoding, flush: true);
+}
+
+/// Saves a [file] from [data] bytes.
+Future<File> saveFileBytes(File file, Uint8List data) async {
+  return file.writeAsBytes(data, flush: true);
 }